game list should be usable now
